+++ src/editors/subscription/goose/goose-publisher-list.ts
@@ -11,11 +11,11 @@ import { translate } from 'lit-translate';
 import '@material/mwc-icon';
 import '@material/mwc-list/mwc-list-item';
 
-import '../../filtered-list.js';
-import { getNameAttribute, newWizardEvent } from '../../foundation.js';
+import '../../../filtered-list.js';
+import { getNameAttribute, newWizardEvent } from '../../../foundation.js';
 import { newGOOSESelectEvent } from './foundation.js';
-import { gooseIcon } from '../../icons/icons.js';
-import { wizards } from '../../wizards/wizard-library.js';
+import { gooseIcon } from '../../../icons/icons.js';
+import { wizards } from '../../../wizards/wizard-library.js';
 import { classMap } from 'lit-html/directives/class-map';
 import { getOrderedIeds, styles } from '../foundation.js';
