+++ test/unit/editors/subscription/goose-publisher-list.test.ts
@@ -1,8 +1,8 @@
 import { html, fixture, expect } from '@open-wc/testing';
 import { SinonSpy, spy } from 'sinon';
 
-import '../../../../src/editors/subscription/goose-publisher-list.js'
-import { GoosePublisherList } from '../../../../src/editors/subscription/goose-publisher-list.js';
+import '../../../../src/editors/subscription/goose/goose-publisher-list.js'
+import { GoosePublisherList } from '../../../../src/editors/subscription/goose/goose-publisher-list.js';
 
 describe('goose-publisher-list', () => {
   let element: GoosePublisherList;
