+++ test/unit/editors/subscription/subscriber-list.test.ts
@@ -1,7 +1,7 @@
 import { html, fixture, expect } from '@open-wc/testing';
 
-import '../../../../src/editors/subscription/subscriber-list.js'
-import { SubscriberList } from '../../../../src/editors/subscription/subscriber-list.js';
+import '../../../../src/editors/subscription/goose/subscriber-list.js'
+import { SubscriberList } from '../../../../src/editors/subscription/goose/subscriber-list.js';
 
 describe('subscriber-list', () => {
   let element: SubscriberList;
